requirements.txt
streamlit
psycopg2-binary
python-dotenv
netwrokx
matplotlib
pyvis
jinja2
watchdog